Refactored code and added pytest unit tests
