+++ data/example_python.json
@@ -1,0 +1,1 @@
+[{"id":1,"name":"Fred"},{"id":2,"name":"Daphne"},{"id":3,"name":"Velma"},{"id":4,"name":"Shaggy"},{"id":5,"name":"Scooby"}]
