+++ data/people_unknown_place.csv
@@ -1,0 +1,14 @@
+given_name,family_name,date_of_birth,place_of_birth
+John,Williams,1842-09-30,Dumfries
+Grace,Jeffery,1899-06-14,Kelso
+Sean,Molnar,1982-11-01,Dromore
+Lily,Doyle,1883-04-02,Hamilton
+Edith,Styles,1879-07-24,Ballymoney
+John,Brown,1892-09-05,Kilmarnock
+Jose,Barth,1942-08-13,Newtownards
+Leonard,Beetham,1897-05-25,Enniskillen
+Christopher,Cash,1977-09-23,Stirling
+Amos,Rogerson,1914-09-12,Jedburgh
+Susan,Helm,1955-04-03,Dromore
+Philip,Venters,1958-05-31,Elgin
+Margaret,Hyam,1945-05-26,Unknown
