+++ data/summary_output.json
@@ -1,0 +1,1 @@
+[{"Scotland":8048},{"Northern Ireland":1952}]
