+++ images/robward/Dockerfile
@@ -6,3 +6,5 @@ COPY requirements.txt ./
 RUN pip install --no-cache-dir -r requirements.txt
 
 COPY example.py ./
+COPY dataload.py ./
+COPY test_dataload.py ./
