+++ images/robward/requirements.txt
@@ -1,2 +1,3 @@
 mysqlclient ~= 2.1.0
 SQLAlchemy ~= 1.4.31
+pytest ~= 6.2.5
